requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
httpx==0.27.2
python-dotenv==1.0.1
fitparse==1.2.0
python-multipart==0.0.12
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
